+++ lib42/ft_sdiv.c
@@ -6,7 +6,7 @@
 /*   By: alambert <alambert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/03 10:57:26 by alambert          #+#    #+#             */
-/*   Updated: 2022/06/05 17:41:38 by alambert         ###   ########.fr       */
+/*   Updated: 2022/06/06 20:50:17 by alambert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,9 +64,9 @@ double	ft_sdd(double a, double b)
 		return (a / b);
 }
 
-float	ft_sdf(float a, float b)
+long double	ft_sdf(long double a, long double b)
 {
-	float	ret;
+	long double	ret;
 
 	if ((a / b) > -FLT_MIN && (a / b) < FLT_MIN && (a / b) != 0)
 	{
